\name{beetle_distM}
\alias{beetle_distM}
\docType{data}
\title{
Gower distance for beetles incidence raw data
}
\description{
beetle_distM is a species-by-species pairwise, symmetric data, which contains all species in beetles incidence raw data. It is used as example for computing functional diversity. \cr
}
\usage{data(Brazil_tree)}
\format{
  beetle_distM is a 179 x 179 (species by species) symmetric data.matrix. The value of beetle_distM is between zero (identical) and one.  \cr
 $ Abdera_flexuosa               : num  0 0.475 0.522 0.176 0.5 ... \cr
 $ Onyxacalles_pyrenaeus         : num  0.475 0 0.129 0.349 0.112 ... \cr
 $ Acanthocinus_griseus          : num  0.5217 0.1286 0 0.3559 0.0392 ... \cr
 ....... \cr
 }
\references{
Thorn, S., Bassler, C., Gottschalk, T., Hothorn, T., Bussler, H., Raffa, K., & Muller, J. (2014). New insights into the consequences of post-windthrow salvage logging revealed by functional structure of saproxylic beetles assemblages. PloS one, 9(7), e101757.
}

\keyword{datasets}
