\name{beetle_tree}
\alias{beetle_tree}
\docType{data}
\title{
Phylogenetic tree for beetles incidence raw data
}
\description{
beetle_tree is a phylogenetic tree, which contains all species in beetles data. It is used as example for computing phylogenetic diversity. \cr
}
\usage{data(beetle_tree)}
\format{
  beetle_tree is a list (phylo tree) with five informations: \cr
 A list of 5: \cr
 $ edge       : int [1:319, 1:2] 180 181 181 180 182 183 184 184 183 182 ... \cr
 $ edge.length: num [1:319] 53.6 214.3 214.3 17.1 100.3 ... \cr
 $ Nnode      : int 141 \cr
 $ node.label : chr [1:141] "match4" "match6" "match35" "match39" ... \cr
 $ tip.label  : chr [1:179] "Plegaderus_vulneratus" "Sphaerites_glabratus" \cr
 }
\references{
Thorn, S., Bassler, C., Gottschalk, T., Hothorn, T., Bussler, H., Raffa, K., & Muller, J. (2014). New insights into the consequences of post-windthrow salvage logging revealed by functional structure of saproxylic beetles assemblages. PloS one, 9(7), e101757.
}

\keyword{datasets}
