% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{DataInfobeta}
\alias{DataInfobeta}
\title{Data information for beta diversity}
\usage{
DataInfobeta(
  data,
  diversity = "TD",
  datatype = "abundance",
  PDtree = NULL,
  PDreftime = NULL,
  FDdistM = NULL,
  FDtype = "AUC",
  FDtau = NULL
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a \code{matrix/data.frame} (species by assemblages), or a \code{list} of \code{matrices/data.frames}, each matrix represents species-by-assemblages abundance matrix.\cr
(b) For \code{datatype = "incidence_raw"}, data can be input as a \code{list} (a dataset) with several \code{lists} (assemblages) of \code{matrices/data.frames}, each matrix represents species-by-sampling units.}

\item{diversity}{selection of diversity type: \code{'TD'} = Taxonomic diversity, \code{'PD'} = Phylogenetic diversity, and \code{'FD'} = Functional diversity.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}) with all entries being \code{0} (non-detection) or \code{1} (detection).}

\item{PDtree}{(required only when \code{diversity = "PD"}), a phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{PDreftime}{(required only when \code{diversity = "PD"}), a numerical value specifying reference time for PD. Default is \code{NULL} (i.e., the age of the root of PDtree).}

\item{FDdistM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species in the pooled assemblage.}

\item{FDtype}{(required only when \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_value"} for FD under a specified threshold value, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{(required only when \code{diversity = "FD"} and \code{FDtype = "tau_value"}), a numerical value between 0 and 1 specifying the tau value (threshold level). If \code{NULL} (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled assemblage (i.e., quadratic entropy).}
}
\value{
return a data.frame.
}
\description{
\code{DataInfobeta}: Data information for beta diversity in three dimension
}
\examples{
## Taxonomic diversity for abundance data
data(Brazil_data)
output1 = DataInfobeta(data = Brazil_data[c('Marim', 'Rebio2')], diversity = 'TD', datatype = 'abundance')
output1


## Taxonomic diversity for incidence data
data(SG_forests)
data = list("CR 2005 vs. 2017" = SG_forests$CR$Year_2005_and_2017,
            "JE 2005 vs. 2017" = SG_forests$JE$Year_2005_and_2017)
output2 = DataInfobeta(data = data, diversity = 'TD', datatype = 'incidence_raw')
output2


## Phylogenetic diversity for abundance data
data(Brazil_PDFD_data)
data(Brazil_tree)
output3 = DataInfobeta(data = Brazil_PDFD_data[c('Marim', 'Rebio2')], diversity = 'PD', datatype = 'abundance', 
                       PDtree = Brazil_tree, PDreftime = NULL)
output3


## Phylogenetic diversity for incidence data
data(beetle_inc)
data(beetle_tree)
output4 = DataInfobeta(data = beetle_inc, diversity = 'PD', datatype = 'incidence_raw', 
                       PDtree = beetle_tree, PDreftime = NULL)
output4


## Functional diversity for abundance data under single threshold
data(Brazil_PDFD_data)
data(Brazil_distM)
output5 = DataInfobeta(data = Brazil_PDFD_data[c('Marim', 'Rebio2')], diversity = 'FD', datatype = 'abundance', 
                       FDdistM = Brazil_distM, FDtype = 'tau_value', FDtau = NULL)
output5


## Functional diversity for incidence data under single threshold
data(beetle_inc)
data(beetle_distM)
output6 = DataInfobeta(data = beetle_inc, diversity = 'FD', datatype = 'incidence_raw', 
                       FDdistM = beetle_distM, FDtype = 'tau_value', FDtau = NULL)
output6


## Functional diversity for abundance data with thresholds integrating from 0 to 1
data(Brazil_PDFD_data)
data(Brazil_distM)
output7 = DataInfobeta(data = Brazil_PDFD_data[c('Marim', 'Rebio2')], diversity = 'FD', datatype = 'abundance', 
                       FDdistM = Brazil_distM, FDtype = 'AUC')
output7


## Functional diversity for incidence data with thresholds integrating from 0 to 1
data(beetle_inc)
data(beetle_distM)
output8 = DataInfobeta(data = beetle_inc, diversity = 'FD', datatype = 'incidence_raw', 
                       FDdistM = beetle_distM, FDtype = 'AUC')
output8
}
