% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggobsbeta3D}
\alias{ggobsbeta3D}
\title{ggplot for beta diversity}
\usage{
ggobsbeta3D(output, type = "B")
}
\arguments{
\item{output}{the output from the function obsbeta3D}

\item{type}{\code{type = 'B'} for plotting gamma, alpha, and beta diversity; \code{type = 'D'} for plotting four dissimilarity indices.}
}
\value{
a figure for gamma, alpha, and Beta diversity, or a figure for four dissimilarity indices.
}
\description{
\code{ggobsbeta3D}: ggplot for observed beta diversity with order q
}
\examples{
## Taxonomic diversity for abundance data
data(Brazil_rainforests)
output1 = obsbeta3D(data = Brazil_rainforests, diversity = 'TD', 
                    datatype = 'abundance', nboot = 50, conf = 0.95)

ggobsbeta3D(output1, type = 'B')
ggobsbeta3D(output1, type = 'D')


## Phylogenetic diversity for abundance data
data(Brazil_rainforests)
data(beetle_tree)
output2 = obsbeta3D(data = Brazil_rainforests, diversity = 'PD', 
                    datatype = 'abundance', nboot = 50, conf = 0.95,
                    PDtree = Brazil_tree, PDreftime = NULL)

ggobsbeta3D(output2, type = 'B')
ggobsbeta3D(output2, type = 'D')


## Functional diversity for abundance data under a specified threshold level
data(Brazil_rainforests)
data(Brazil_distM)
output3 = obsbeta3D(data = Brazil_rainforests, diversity = 'FD', 
                    datatype = 'abundance', nboot = 50, conf = 0.95, 
                    FDdistM = Brazil_distM, FDtype = 'tau_value', FDtau = NULL)

ggobsbeta3D(output3, type = 'B')
ggobsbeta3D(output3, type = 'D')


## Functional diversity for abundance data when all thresholds from 0 to 1 are considered
data(Brazil_rainforests)
data(Brazil_distM)
output4 = obsbeta3D(data = Brazil_rainforests, diversity = 'FD', 
                    datatype = 'abundance', nboot = 50, conf = 0.95, 
                    FDdistM = Brazil_distM, FDtype = 'AUC', FDcut_number = 50)

ggobsbeta3D(output4, type = 'B')
ggobsbeta3D(output4, type = 'D')


## Taxonomic diversity for incidence data
data(Second_growth_forests)
output5 = obsbeta3D(data = Second_growth_forests, diversity = 'TD', 
                    datatype = 'incidence_raw', nboot = 50, conf = 0.95)

ggobsbeta3D(output5, type = 'B')
ggobsbeta3D(output5, type = 'D')


}
