% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{iNEXTbeta3D_pair}
\alias{iNEXTbeta3D_pair}
\title{Dissimilarity estimates for all pairs of assemblages}
\usage{
iNEXTbeta3D_pair(data, SC = NULL, parallel = F, cpus = 2)
}
\arguments{
\item{data}{a data.frame/matrix with species (rows) by assemblages (columns).}

\item{SC}{a standardized sample coverage value for computing four standardized (estimated) dissimilarity indices. 
By default (\code{SC = NULL}), based on all pairs of samples, SC represents the minimum among the coverage values for alpha reference samples
extrapolated to double the total abundance/size.  
The default coverage value is referred to as C_2n,alpha in Chao et al. (2023, Ecological Monographs) or Cmax_joint in this code; 
it represents the maximum coverage value for which one can reliably infer dissimilarity and beta diversity. This value varies with dataset. 
When there are many datasets, a common value of SC should be specified for the argument(e.g., (\code{SC=0.9})).}

\item{parallel}{whether to do parallel computation or not.}

\item{cpus}{number of cpu cores for parallel.}
}
\value{
a data frame including observed and standardized Jaccard dissimilarity (\code{jac_obs} and \code{jac_est}), observed and standardized Sorensen dissimilarity (\code{sor_obs} and \code{sor_est}), 
observed and standardized Horn index (\code{hor_obs} and \code{hor_est}), as well as observed and standardized Morisita-Horn index (\code{mor_hor_obs} and \code{mor_hor_est}) for all pairs of assemblages
under the default or user-specified sample coverage (SC).
}
\description{
\code{iNEXTbeta3D_pair} compute observed and coverage-based standardized (estimated) Jaccard dissimilarity index (q=0), Sorensen dissimilarity index (q=0),
Horn dissimilarity index (q=1), and Morisita-Horn dissimilarity index (q=2) for all pairs of assemblages.
NOTE: The observed dissimilarity indices typically underestimates; they are computed only for comparison.
We suggest researchers use the coverage-based standardized dissimilarity indices.
}
