% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{obsbeta3D}
\alias{obsbeta3D}
\title{Observed beta diversity with order q for TD, PD, and FD}
\usage{
obsbeta3D(
  data,
  diversity = "TD",
  q = seq(0, 2, 0.25),
  datatype = "abundance",
  nboot = 20,
  conf = 0.95,
  PDtree = NULL,
  PDreftime = NULL,
  PDtype = "meanPD",
  FDdistM = NULL,
  FDtype = "AUC",
  FDtau = NULL,
  FDcut_number = 30
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, species abundance data for a single dataset can be input as a \code{matrix/data.frame} (species-by-assemblage); data for multiple datasets can be input as a \code{list} of \code{matrices/data.frames}, with each matrix representing a species-by-assemblage abundance matrix for one of the datasets.\cr
(b) For \code{datatype = "incidence_raw"}, data for a single dataset with N assemblages can be input as a \code{list} of \code{matrices/data.frames}, with each matrix representing a species-by-sampling-unit incidence matrix for one of the assemblages; data for multiple datasets can be input as multiple lists.}

\item{diversity}{selection of diversity type: \code{'TD'} = Taxonomic diversity, \code{'PD'} = Phylogenetic diversity, and \code{'FD'} = Functional diversity.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{q = c(0, 1, 2)}.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}) with all entries being \code{0} (non-detection) or \code{1} (detection).}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter \code{0} to skip the bootstrap procedures. Default is \code{nboot = 10}. If more accurate results are required, set \code{nboot = 100} (or \code{200}).}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{conf = 0.95}.}

\item{PDtree}{(required argument for \code{diversity = "PD"}), a phylogenetic tree in Newick format for all observed species in the pooled dataset.}

\item{PDreftime}{(argument for \code{diversity = "PD"}), a numerical value specifying reference time for PD. Default is \code{PDreftime = NULL} (i.e., the age of the root of PDtree).}

\item{PDtype}{(argument for \code{diversity = "PD"}), select PD type: \code{PDtype = "PD"} (effective total branch length) or \code{PDtype = "meanPD"} (effective number of equally divergent lineages). Default is \code{PDtype = "meanPD"}, where \code{meanPD = PD/tree depth}.}

\item{FDdistM}{(required argument for \code{diversity = "FD"}), a species pairwise distance matrix for all species in the pooled dataset.}

\item{FDtype}{(argument for \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_value"} for FD under a specified threshold value, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is \code{FDtype = "AUC"}.}

\item{FDtau}{(argument for \code{diversity = "FD"} and \code{FDtype = "tau_value"}), a numerical value between 0 and
1 specifying the tau value (threshold level) that will be used to compute FD. If \code{FDtype = NULL} (default), 
then threshold level is set to be the mean distance between any two individuals randomly selected from the pooled 
dataset (i.e., quadratic entropy).}

\item{FDcut_number}{(argument for \code{diversity = "FD"} and \code{FDtype = "AUC"}), a numeric number to cut [0, 1] interval into equal-spaced sub-intervals to obtain the AUC value by integrating the tau-profile. Equivalently, the number of tau values that will be considered to compute the integrated AUC value. Default is \code{FDcut_number = 30}. A larger value can be set to obtain more accurate AUC value.}
}
\value{
Return a list of several data frames including three diversity (gamma, alpha, and beta
 diversity) and four dissimilarity measures.\cr 
 
 The output in each data frame includes: 
 \item{Dataset}{the name of dataset.}
 \item{Order.q}{the diversity order of q.} 
 \item{Size}{the sample size.}
 \item{Value}{the gamma, alpha, beta diversity and dissimilarity value.}
 \item{s.e.}{standard error of standardized estimate.}
 \item{LCL, UCL}{the bootstrap lower and upper confidence limits for the diversity/dissimilarity with a default significance level of 0.95.}
 \item{Diversity}{'TD' = 'Taxonomic diversity', 'PD' = 'Phylogenetic diversity', 'meanPD' = 'Mean phylogenetic diversity', 'FD_tau' = 'Functional diversity (given tau)', 'FD_AUC' = 'Functional diversity (AUC)'}
 \item{Type}{gamma, alpha, beta diversity or dissimilarity measures.}
}
\description{
\code{obsbeta3D}: compute observed 3D alpha, beta, gamma diversity as well as dissimilarity indices.
}
\examples{
## Taxonomic diversity for abundance data
data(Brazil_rainforests)
output1 = obsbeta3D(data = Brazil_rainforests, diversity = 'TD', 
                    datatype = 'abundance', nboot = 50, conf = 0.95)
output1


## Phylogenetic diversity for abundance data
data(Brazil_rainforests)
data(Brazil_tree)
output2 = obsbeta3D(data = Brazil_rainforests, diversity = 'PD', 
                    datatype = 'abundance', nboot = 50, conf = 0.95,
                    PDtree = Brazil_tree, PDreftime = NULL)
output2


## Functional diversity for abundance data under a specified threshold level
data(Brazil_rainforests)
data(Brazil_distM)
output3 = obsbeta3D(data = Brazil_rainforests, diversity = 'FD', 
                    datatype = 'abundance', nboot = 50, conf = 0.95, 
                    FDdistM = Brazil_distM, FDtype = 'tau_value', FDtau = NULL)
output3


## Functional diversity for abundance data when all thresholds from 0 to 1 are considered
data(Brazil_rainforests)
data(Brazil_distM)
output4 = obsbeta3D(data = Brazil_rainforests, diversity = 'FD', 
                    datatype = 'abundance', nboot = 50, conf = 0.95, 
                    FDdistM = Brazil_distM, FDtype = 'AUC', FDcut_number = 50)
output4


## Taxonomic diversity for incidence data
data(Second_growth_forests)
output5 = obsbeta3D(data = Second_growth_forests, diversity = 'TD', 
                    datatype = 'incidence_raw', nboot = 50, conf = 0.95)
output5

}
