% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{print.iNEXT.beta3D}
\alias{print.iNEXT.beta3D}
\title{Printing iNEXT.beta3D object}
\usage{
\method{print}{iNEXT.beta3D}(x, ...)
}
\arguments{
\item{x}{an \code{iNEXT.beta3D} object computed by \code{\link{iNEXTbeta3D}}.}

\item{...}{additional arguments.}
}
\description{
\code{print.iNEXT.beta3D}: Print method for objects inheriting from class "iNEXT.beta3D"
}
