% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{print.iNEXTbeta3D}
\alias{print.iNEXTbeta3D}
\title{Printing iNEXTbeta3D object}
\usage{
\method{print}{iNEXTbeta3D}(x, ...)
}
\arguments{
\item{x}{an \code{iNEXTbeta3D} object computed by \code{\link{iNEXTbeta3D}}.}

\item{...}{additional arguments.}
}
\description{
\code{print.iNEXTbeta3D}: Print method for objects inheriting from class "iNEXTbeta3D"
}
