\name{Brazil_distM}
\alias{Brazil_distM}
\docType{data}
\title{
Gower distance for Brazil abundance data
}
\description{
Brazil_distM is a species-by-species pairwise, symmetric data, which contains all species collected from eight rainforest fragments in Brazil_rainforests data. It is used as example for computing functional diversity. \cr
}
\usage{data(Brazil_tree)}
\format{
  Brazil_distM is a 425 x 425 (species by species) symmetric data.frame. The value of Brazil_distM is between zero (identical) and one.  \cr
 $ Carpotroche_brasiliensis      : num  0 0.522 0.522 0.253 0.263 ... \cr
 $ Astronium_concinnum           : num  0.522 0 0 0.525 0.509 ... \cr
 $ Astronium_graveolens          : num  0.522 0 0 0.525 0.509 ... \cr
 ....... \cr
 }
\references{
Magnago, L. F. S., Edwards, D. P., Edwards, F. A., Magrach, A., Martins, S. V., & Laurance, W. F. 2014. Functional attributes change but functional richness is unchanged after fragmentation of Brazilian Atlantic forests. Journal of Ecology 102: 475-485. \cr\cr
Magnago, L. F. S., Magrach, A., Barlow, J., Schaefer, C. E. G. R., Laurance, W. F., Martins, S. V., & Edwards, D. P. 2017. Do fragment size and edge effects predict carbon stocks in trees and lianas in tropical forests? Functional Ecology 31: 542-552. \cr\cr
Magnago, L. F. S., Rocha, M. F., Meyer, L., Martins, S. v., & Meira-Neto, J. A. A. 2015. Microclimatic conditions at forest edges have significant impacts on vegetation structure in large Atlantic forest fragments. Biodiversity and Conservation 24: 2305-2318. \cr\cr 
}

\keyword{datasets}
