\name{Brazil_tree}
\alias{Brazil_tree}
\docType{data}
\title{
  Phylogenetic tree for Brazil abundance data
}
\description{
  This dataset includes a phylogenetic tree spanned by 185 species listed in the \code{Brazil_rainforests} dataset.  \cr
}
\usage{data(Brazil_tree)}
\format{
  \code{Brazil_tree} is a list (phylo tree) with the following phylogenetic information: \cr
  A list of 5: \cr
  $ edge       : int [1:301, 1:2] 186 187 188 189 190 191 192 193 193 194 ... \cr
  $ edge.length: num [1:301] 75 146 8 6 18 10 10 127 6 23 ... \cr
  $ Nnode      : int 117 \cr
  $ node.label : chr [1:117] "magnoliales_to_asterales" "poales_to_asterales" ... \cr
  $ tip.label  : chr [1:185] "Carpotroche_brasiliensis" "Casearia_ulmifolia" "Casearia_sp2" ... \cr
}

\keyword{datasets}
