\name{Brazil_tree}
\alias{Brazil_tree}
\docType{data}
\title{
  Phylogenetic tree for Brazil abundance data
}
\description{
  This dataset includes a phylogenetic tree spanned by 243 species listed in the \code{Brazil_rainforests} dataset.  \cr
}
\usage{data(Brazil_tree)}
\format{
  \code{Brazil_tree} is a list (phylo tree) with the following phylogenetic information: \cr
  A list of 5: \cr
  $ edge       : int [1:382, 1:2] 244 245 246 247 248 249 250 251 251 252 ... \cr
  $ edge.length: num [1:382] 75 146 8 6 18 10 10 127 6 23 ... \cr
  $ Nnode      : int 140 \cr
  $ node.label : chr [1:140] "magnoliales_to_asterales" "poales_to_asterales" ... \cr
  $ tip.label  : chr [1:243] "Carpotroche_brasiliensis" "Casearia_ulmifolia" "Casearia_sp4" ... \cr
}

\keyword{datasets}
