\name{Brazil_tree}
\alias{Brazil_tree}
\docType{data}
\title{
Phylogenetic tree for Brazil abundance data
}
\description{
Brazil_tree is a phylogenetic tree, which contains all species collected from rainforest fragments in Brazil. It is used as example for computing phylogenetic diversity. \cr
}
\usage{data(Brazil_tree)}
\format{
  Brazil_tree is a list (phylo tree) with five informations: \cr
 A list of 5: \cr
 $ edge       : int [1:629, 1:2] 426 427 428 429 430 431 432 433 434 434 ... \cr
 $ edge.length: num [1:629] 75 146 8 6 4 14 10 10 127 6 ... \cr
 $ Nnode      : int 205 \cr
 $ node.label : chr [1:205] "magnoliales_to_asterales" "poales_to_asterales", ... \cr
 $ tip.label  : chr [1:425] "Carpotroche_brasiliensis" "Casearia_ulmifolia" "Casearia_sp4", ... \cr
 }
\references{
Magnago, L. F. S., Edwards, D. P., Edwards, F. A., Magrach, A., Martins, S. V., & Laurance, W. F. 2014. Functional attributes change but functional richness is unchanged after fragmentation of Brazilian Atlantic forests. Journal of Ecology 102: 475-485. \cr\cr
Magnago, L. F. S., Magrach, A., Barlow, J., Schaefer, C. E. G. R., Laurance, W. F., Martins, S. V., & Edwards, D. P. 2017. Do fragment size and edge effects predict carbon stocks in trees and lianas in tropical forests? Functional Ecology 31: 542-552. \cr\cr
Magnago, L. F. S., Rocha, M. F., Meyer, L., Martins, S. v., & Meira-Neto, J. A. A. 2015. Microclimatic conditions at forest edges have significant impacts on vegetation structure in large Atlantic forest fragments. Biodiversity and Conservation 24: 2305-2318. \cr\cr
}

\keyword{datasets}
