% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggiNEXTbeta3D}
\alias{ggiNEXTbeta3D}
\title{ggplot2 extension for an iNEXTbeta3D object}
\usage{
ggiNEXTbeta3D(output, type = "B", scale = "free", transp = 0.4)
}
\arguments{
\item{output}{the output from iNEXTbeta3D}

\item{type}{(required only when \code{base = "coverage"}), selection of plot type : \cr
\code{type = 'B'} for plotting the gamma, alpha, and beta diversity ;  \cr
\code{type = 'D'} for plotting 4 turnover dissimilarities.}

\item{scale}{Are scales shared across all facets (the default, \code{"fixed"}), or do they vary across rows (\code{"free_x"}), columns (\code{"free_y"}), or both rows and columns (\code{"free"})? Default is \code{"free"}.}

\item{transp}{a value between 0 and 1 controlling transparency. \code{transp = 0} is completely transparent, default is 0.4.}
}
\value{
a figure for Beta diversity or dissimilarity index.
}
\description{
\code{ggiNEXTbeta3D}: the \code{\link[ggplot2]{ggplot}} extension for \code{\link{iNEXTbeta3D}} object to plot sample-size- and coverage-based rarefaction/extrapolation curves.
}
\examples{
## Taxonomic diversity for abundance data
data(beetle_abu)
output1 = iNEXTbeta3D(data = beetle_abu, diversity = 'TD', datatype = 'abundance', 
                      level = seq(0.8, 1, 0.05), nboot = 20)

ggiNEXTbeta3D(output1, type = 'B', scale = 'free')
ggiNEXTbeta3D(output1, type = 'D', scale = 'free')


## Taxonomic diversity for incidence data
data(beetle_inc)
output2 = iNEXTbeta3D(data = beetle_inc, diversity = 'TD', datatype = 'incidence_raw', 
                      level = seq(0.8, 1, 0.05), nboot = 20, conf = 0.95)

ggiNEXTbeta3D(output2, type = 'B', scale = 'free')
ggiNEXTbeta3D(output2, type = 'D', scale = 'free')


## Phylogenetic diversity for abundance data
data(beetle_abu)
data(beetle_tree)
output3 = iNEXTbeta3D(data = beetle_abu, diversity = 'PD', datatype = 'abundance', 
                      level = seq(0.8, 1, 0.05), nboot = 20, conf = 0.95, 
                      PDtree = beetle_tree, PDreftime = NULL, PDtype = 'meanPD')

ggiNEXTbeta3D(output3, type = 'B', scale = 'free')
ggiNEXTbeta3D(output3, type = 'D', scale = 'free')


## Phylogenetic diversity for incidence data
data(beetle_inc)
data(beetle_tree)
output4 = iNEXTbeta3D(data = beetle_inc, diversity = 'PD', datatype = 'incidence_raw', 
                      level = seq(0.8, 1, 0.05), nboot = 0, conf = 0.95, 
                      PDtree = beetle_tree, PDreftime = NULL, PDtype = 'meanPD')

ggiNEXTbeta3D(output4, type = 'B', scale = 'free')
ggiNEXTbeta3D(output4, type = 'D', scale = 'free')


## Functional diversity for abundance data under single threshold
data(beetle_abu)
data(beetle_distM)
output5 = iNEXTbeta3D(data = beetle_abu, diversity = 'FD', datatype = 'abundance', 
                      level = seq(0.8, 1, 0.05), nboot = 20, conf = 0.95, 
                      FDdistM = beetle_distM, FDtype = 'tau_value', FDtau = NULL)

ggiNEXTbeta3D(output5, type = 'B', scale = 'free')
ggiNEXTbeta3D(output5, type = 'D', scale = 'free')


## Functional diversity for incidence data under single threshold
data(beetle_inc)
data(beetle_distM)
output6 = iNEXTbeta3D(data = beetle_inc, diversity = 'FD', datatype = 'incidence_raw', 
                      level = seq(0.8, 1, 0.05), nboot = 20, conf = 0.95, 
                      FDdistM = beetle_distM, FDtype = 'tau_value', FDtau = NULL)

ggiNEXTbeta3D(output6, type = 'B', scale = 'free')
ggiNEXTbeta3D(output6, type = 'D', scale = 'free')


## Functional diversity for abundance data with thresholds integrating from 0 to 1
data(beetle_abu)
data(beetle_distM)
output7 = iNEXTbeta3D(data = beetle_abu, diversity = 'FD', datatype = 'abundance', 
                      level = seq(0.8, 1, 0.05), nboot = 10, conf = 0.95, 
                      FDdistM = beetle_distM, FDtype = 'AUC', FDcut_number = 30)

ggiNEXTbeta3D(output7, type = 'B', scale = 'free')
ggiNEXTbeta3D(output7, type = 'D', scale = 'free')


## Functional diversity for incidence data with thresholds integrating from 0 to 1
data(beetle_inc)
data(beetle_distM)
output8 = iNEXTbeta3D(data = beetle_inc, diversity = 'FD', datatype = 'incidence_raw', 
                      level = seq(0.8, 1, 0.05), nboot = 10, conf = 0.95, 
                      FDdistM = beetle_distM, FDtype = 'AUC', FDcut_number = 30)

ggiNEXTbeta3D(output8, type = 'B', scale = 'free')
ggiNEXTbeta3D(output8, type = 'D', scale = 'free')
}
