% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggiNEXTbeta3D}
\alias{ggiNEXTbeta3D}
\title{ggplot2 extension for an iNEXTbeta3D object}
\usage{
ggiNEXTbeta3D(output, type = "B")
}
\arguments{
\item{output}{the output from function iNEXTbeta3D}

\item{type}{(argument only when \code{base = "coverage"}), selection of plot type : \cr
\code{type = 'B'} for plotting the gamma, alpha, and beta diversity ;  \cr
\code{type = 'D'} for plotting 4 turnover dissimilarities.}
}
\value{
a figure for gamma, alpha, and beta diversity or four dissimilarity measures when \code{base = "coverage"}; or a figure for gamma, alpha diversity when \code{base = "size"}.\cr
}
\description{
\code{ggiNEXTbeta3D}: the \code{\link[ggplot2]{ggplot}} extension for \code{\link{iNEXTbeta3D}} object to plot sample-size- and coverage-based rarefaction/extrapolation curves.
}
\examples{
## Taxonomic diversity for abundance data
# Coverage-based
data(Brazil_rainforests)
output1c = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'TD', 
                       datatype = 'abundance', base = "coverage", nboot = 30, conf = 0.95)

ggiNEXTbeta3D(output1c, type = 'B')
ggiNEXTbeta3D(output1c, type = 'D')
 

# Size-based
data(Brazil_rainforests)
output1s = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'TD', 
                       datatype = 'abundance', base = "size", nboot = 30, conf = 0.95)

ggiNEXTbeta3D(output1s)


## Phylogenetic diversity for abundance data
# Coverage-based
data(Brazil_rainforests)
data(Brazil_tree)
output2c = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'PD', 
                       datatype = 'abundance', base = "coverage", nboot = 10, conf = 0.95, 
                       PDtree = Brazil_tree, PDreftime = NULL, PDtype = 'PD')

ggiNEXTbeta3D(output2c, type = 'B')
ggiNEXTbeta3D(output2c, type = 'D')


# Size-based
data(Brazil_rainforests)
data(Brazil_tree)
output2s = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'PD', 
                       datatype = 'abundance', base = "size", nboot = 10, conf = 0.95, 
                       PDtree = Brazil_tree, PDreftime = NULL, PDtype = 'PD')

ggiNEXTbeta3D(output2s)


## Functional diversity for abundance data under single threshold
# Coverage-based
data(Brazil_rainforests)
data(Brazil_distM)
output3c = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'FD', 
                       datatype = 'abundance', base = "coverage", nboot = 30, conf = 0.95, 
                       FDdistM = Brazil_distM, FDtype = 'tau_value', FDtau = NULL)

ggiNEXTbeta3D(output3c, type = 'B')
ggiNEXTbeta3D(output3c, type = 'D')


# Size-based
data(Brazil_rainforests)
data(Brazil_distM)
output3s = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'FD', 
                       datatype = 'abundance', base = "size", nboot = 30, conf = 0.95, 
                       FDdistM = Brazil_distM, FDtype = 'tau_value', FDtau = NULL)

ggiNEXTbeta3D(output3s)


## Functional diversity for abundance data with thresholds integrating from 0 to 1
# Coverage-based
data(Brazil_rainforests)
data(Brazil_distM)
output4c = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'FD', 
                       datatype = 'abundance', base = "coverage", nboot = 0, conf = 0.95, 
                       FDdistM = Brazil_distM, FDtype = 'AUC', FDcut_number = 30)

ggiNEXTbeta3D(output4c, type = 'B')
ggiNEXTbeta3D(output4c, type = 'D')


# Size-based
data(Brazil_rainforests)
data(Brazil_distM)
output4s = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'FD', 
                       datatype = 'abundance', base = "size", nboot = 10, conf = 0.95, 
                       FDdistM = Brazil_distM, FDtype = 'AUC', FDcut_number = 30)

ggiNEXTbeta3D(output4s)


## Taxonomic diversity for incidence data
# Coverage-based
data(Second_growth_forests)
output5c = iNEXTbeta3D(data = Second_growth_forests, diversity = 'TD', datatype = 'incidence_raw', 
                       level = NULL, nboot = 20, conf = 0.95)

ggiNEXTbeta3D(output5c, type = 'B')
ggiNEXTbeta3D(output5c, type = 'D')


# Size-based
data(Second_growth_forests)
output5s = iNEXTbeta3D(data = Second_growth_forests, diversity = 'TD', datatype = 'incidence_raw', 
                       base = "size", level = NULL, nboot = 30, conf = 0.95)

ggiNEXTbeta3D(output5s)


}
\references{
Chao, A., Thorn, S., Chiu, C.-H., Moyes, F., Hu, K.-H., Chazdon, R. L., Wu, J., Dornelas, M., Zelený, D., Colwell, R. K., and Magurran, A. E. (2023). Rarefaction and extrapolation with beta diversity under a framework of Hill numbers: the iNEXT.beta3D standardization. To appear in Ecological Monographs.
}
