% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggiNEXTbeta3D}
\alias{ggiNEXTbeta3D}
\title{ggplot2 extension for the iNEXTbeta3D object}
\usage{
ggiNEXTbeta3D(output, type = "B")
}
\arguments{
\item{output}{output from the function \code{iNEXTbeta3D}.}

\item{type}{(argument only for \code{base = "coverage"}),\cr
\code{type = 'B'} for plotting the rarefaction and extrapolation sampling curves for gamma, alpha, and beta diversity;  \cr
\code{type = 'D'} for plotting the rarefaction and extrapolation sampling curves for four dissimilarity indices.\cr
Skip the argument for plotting size-based rarefaction and extrapolation sampling curves for gamma and alpha diversity.}
}
\value{
A figure showing gamma, alpha, and beta diversity or the four dissimilarity
indices when \code{base = "coverage"}, or a figure showing gamma and alpha diversity
when \code{base = "size"}.
}
\description{
\code{ggiNEXTbeta3D} is an \code{ggplot2} extension for the \code{iNEXTbeta3D} 
object to plot sample-size- and coverage-based rarefaction/extrapolation curves.
(This function only accepts the output from \code{iNEXTbeta3D} with \code{by_pair = FALSE})
}
\examples{
\donttest{
## (Graphic Display) Taxonomic diversity for abundance data
# Coverage-based rarefaction and extrapolation sampling curves 
# See Example 1 in the iNEXT.beta3D vignette for details.

data(Brazil_rainforests)
output_TDc_abun = iNEXTbeta3D(data = Brazil_rainforests[1:2], diversity = 'TD', 
                              datatype = 'abundance', base = "coverage", nboot = 10)

ggiNEXTbeta3D(output_TDc_abun, type = 'B')
ggiNEXTbeta3D(output_TDc_abun, type = 'D')
}

# Size-based rarefaction and extrapolation sampling curves 

data(Brazil_rainforests)
output_TDs_abun = iNEXTbeta3D(data = Brazil_rainforests[1:2], diversity = 'TD', 
                              datatype = 'abundance', base = "size", nboot = 10)

ggiNEXTbeta3D(output_TDs_abun)

\donttest{
## (Graphic Display) Taxonomic diversity for incidence data
# Coverage-based rarefaction and extrapolation sampling curves 
# See Example 3 in the iNEXT.beta3D vignette for details.

data(Second_growth_forests)
data = list("CR 2005 vs. 2017" = Second_growth_forests[[1]][c(1,3)],
            "JE 2005 vs. 2017" = Second_growth_forests[[2]][c(1,3)])
output_TDc_inci = iNEXTbeta3D(data = data, diversity = 'TD', 
                              datatype = 'incidence_raw', base = "coverage", nboot = 10)

ggiNEXTbeta3D(output_TDc_inci, type = 'B')
ggiNEXTbeta3D(output_TDc_inci, type = 'D')
}

# Size-based rarefaction and extrapolation sampling curves 

data(Second_growth_forests)
data = list("CR 2005 vs. 2017" = Second_growth_forests[[1]][c(1,3)],
            "JE 2005 vs. 2017" = Second_growth_forests[[2]][c(1,3)])
output_TDs_inci = iNEXTbeta3D(data = data, diversity = 'TD', 
                              datatype = 'incidence_raw', base = "size", nboot = 10)

ggiNEXTbeta3D(output_TDs_inci)

\donttest{
## (Graphic Display) Phylogenetic diversity for abundance data
# Coverage-based rarefaction and extrapolation sampling curves 
# See Example 5 in the iNEXT.beta3D vignette for details.

data(Brazil_rainforests)
data(Brazil_tree)
output_PDc_abun = iNEXTbeta3D(data = Brazil_rainforests[1:2], diversity = 'PD', 
                              datatype = 'abundance', base = "coverage", nboot = 10, 
                              PDtree = Brazil_tree, PDreftime = NULL, PDtype = 'AUC', PDcut_number = 30) 

ggiNEXTbeta3D(output_PDc_abun, type = 'B')
ggiNEXTbeta3D(output_PDc_abun, type = 'D')


# Size-based rarefaction and extrapolation sampling curves 

data(Brazil_rainforests)
data(Brazil_tree)
output_PDs_abun = iNEXTbeta3D(data = Brazil_rainforests[1:2], diversity = 'PD', 
                              datatype = 'abundance', base = "size", nboot = 10, 
                              PDtree = Brazil_tree, PDreftime = NULL, PDtype = 'AUC', PDcut_number = 30) 

ggiNEXTbeta3D(output_PDs_abun)


## (Graphic Display) Functional diversity for abundance data when all threshold levels 
## from 0 to 1 are considered
# Coverage-based rarefaction and extrapolation sampling curves 
# See Example 7 in the iNEXT.beta3D vignette for details.

data(Brazil_rainforests)
data(Brazil_distM)
output_FDc_abun = iNEXTbeta3D(data = Brazil_rainforests[1:2], diversity = 'FD', 
                              datatype = 'abundance', base = "coverage", nboot = 10, 
                              FDdistM = Brazil_distM, FDtype = 'AUC', FDcut_number = 30)

ggiNEXTbeta3D(output_FDc_abun, type = 'B')
ggiNEXTbeta3D(output_FDc_abun, type = 'D')


# Size-based rarefaction and extrapolation sampling curves 

data(Brazil_rainforests)
data(Brazil_distM)
output_FDs_abun = iNEXTbeta3D(data = Brazil_rainforests[1:2], diversity = 'FD', 
                              datatype = 'abundance', base = "size", nboot = 10, 
                              FDdistM = Brazil_distM, FDtype = 'AUC', FDcut_number = 30)

ggiNEXTbeta3D(output_FDs_abun)
}


}
