% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{iNEXTBeta3D}
\alias{iNEXTBeta3D}
\title{iNterpolation and EXTrapolation for Beta diversity}
\usage{
iNEXTBeta3D(
  data,
  diversity = "TD",
  q = c(0, 1, 2),
  datatype = "abundance",
  level,
  nboot = 20,
  conf = 0.95,
  PDtree = NULL,
  PDreftime = NULL,
  PDtype = "meanPD",
  FDdistM = NULL,
  FDtype = "AUC",
  FDtau = NULL,
  FDcut_number = 30
)
}
\arguments{
\item{data}{a \code{matrix}, \code{data.frame} (species by assemblages), or \code{list} of species abundance/incidence raw data.\cr}

\item{diversity}{selection of diversity type: \code{'TD'} = Taxonomic diversity, \code{'PD'} = Phylogenetic diversity, and \code{'FD'} = Functional diversity.}

\item{q}{a numerical vector specifying the diversity orders.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).}

\item{level}{a sequence specifying the particular sample coverages (between 0 and 1). Calculating asymptotic diversity if \code{C}=1. \cr}

\item{nboot}{an integer specifying the number of replications for bootstrapping.}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95.}

\item{PDtree}{a phylo object describing the phylogenetic tree in Newick format for all observed species in the pooled assemblage. It is necessary when \code{diversity = 'PD'}.}

\item{PDreftime}{Select a reference time point for \code{diversity = 'PD'}. Default is set to be the tree depth of the phylogenetic tree, which is spanned by all the observed species in the pooled assemblage. Only required when \code{diversity = 'PD'}.}

\item{PDtype}{Select phylogenetic diversity type: \code{PDtype = "PD"} for Chao et al. (2010) phylogenetic diversity and \code{PDtype = "meanPD"} for mean phylogenetic diversity (phylogenetic Hill number). Only required when \code{diversity = 'PD'}. Default is \code{"meanPD"}.}

\item{FDdistM}{a pair wise distance matrix for all pairs of observed species in the pooled assemblage. It is necessary when diversity = 'FD'.}

\item{FDtype}{a binary selection for FD. \code{FDtype = "tau_value"} computes diversity under a certain threshold value. \code{FDtype = "AUC"} computes an overall FD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{a value between 0 and 1 specifying tau. If \code{NULL}, \code{threshold = } dmean. Default is \code{NULL}. Only required when \code{diversity = 'FD'} and \code{FDtype = "tau_value"}.}

\item{FDcut_number}{a numeric number to split zero to one into several equal-spaced length. Only required when \code{diversity = "FD"} and \code{FDtype = "AUC"}. Default is 30.}
}
\value{
A list of seven lists with three-diversity and four-dissimilarity.
}
\description{
\code{iNEXTBeta3D}: Interpolation and extrapolation of Beta diversity with order q
}
\examples{
## Taxonomic diversity for abundance data
data(beetle_abu)
output1 = iNEXTBeta3D(data = beetle_abu, diversity = 'TD', datatype = 'abundance', level = seq(0.8, 1, 0.05), 
                      nboot = 20, conf = 0.95)
output1


## Taxonomic diversity for incidence data
data(beetle_inc)
output2 = iNEXTBeta3D(data = beetle_inc, diversity = 'TD', datatype = 'incidence_raw', level = seq(0.8, 1, 0.05),
                      nboot = 20, conf = 0.95)
output2


## Phylogenetic diversity for abundance data
data(beetle_abu)
data(beetle_tree)
output3 = iNEXTBeta3D(data = beetle_abu, diversity = 'PD', datatype = 'abundance',  level = seq(0.8, 1, 0.05),
                      nboot = 20, conf = 0.95, PDtree = beetle_tree, PDreftime = NULL, PDtype = 'PD')
output3


## Phylogenetic diversity for incidence data
data(beetle_inc)
data(beetle_tree)
output4 = iNEXTBeta3D(data = beetle_inc, diversity = 'PD', datatype = 'incidence_raw', level = seq(0.8, 1, 0.05),
                      nboot = 20, conf = 0.95, PDtree = beetle_tree, PDreftime = NULL, PDtype = 'PD')
output4


## Functional diversity for abundance data under single threshold
data(beetle_abu)
data(beetle_distM)
output5 = iNEXTBeta3D(data = beetle_abu, diversity = 'FD', datatype = 'abundance', level = seq(0.8, 1, 0.05),
                      nboot = 20, conf = 0.95, FDdistM = beetle_distM, FDtype = 'tau_value', FDtau = NULL)
output5


## Functional diversity for incidence data under single threshold
data(beetle_inc)
data(beetle_distM)
output6 = iNEXTBeta3D(data = beetle_inc, diversity = 'FD', datatype = 'incidence_raw', level = seq(0.8, 1, 0.05),
                      nboot = 20, conf = 0.95, FDdistM = beetle_distM, FDtype = 'tau_value', FDtau = NULL)
output6


## Functional diversity for abundance data with thresholds integrating from 0 to 1
data(beetle_abu)
data(beetle_distM)
output7 = iNEXTBeta3D(data = beetle_abu, diversity = 'FD', datatype = 'abundance', level = seq(0.8, 1, 0.05),
                      nboot = 10, conf = 0.95, FDdistM = beetle_distM, FDtype = 'AUC', FDcut_number = 30)
output7


## Functional diversity for incidence data with thresholds integrating from 0 to 1
data(beetle_inc)
data(beetle_distM)
output8 = iNEXTBeta3D(data = beetle_inc, diversity = 'FD', datatype = 'incidence_raw', level = seq(0.8, 1, 0.05),
                      nboot = 10, conf = 0.95, FDdistM = beetle_distM, FDtype = 'AUC', FDcut_number = 30)
output8

}
