% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{iNEXTbeta3D}
\alias{iNEXTbeta3D}
\title{iNterpolation and EXTrapolation for beta diversity}
\usage{
iNEXTbeta3D(
  data,
  diversity = "TD",
  q = c(0, 1, 2),
  datatype = "abundance",
  base = "coverage",
  level = NULL,
  nboot = 20,
  conf = 0.95,
  PDtree = NULL,
  PDreftime = NULL,
  PDtype = "meanPD",
  FDdistM = NULL,
  FDtype = "AUC",
  FDtau = NULL,
  FDcut_number = 30
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a \code{matrix/data.frame} (species by assemblages), or a \code{list} of \code{matrices/data.frames}, each matrix represents species-by-assemblages abundance matrix.\cr
(b) For \code{datatype = "incidence_raw"}, data can be input as a \code{list} (a region) with several \code{lists} (assemblages) of \code{matrices/data.frames}, each matrix represents species-by-sampling units.}

\item{diversity}{selection of diversity type: \code{'TD'} = Taxonomic diversity, \code{'PD'} = Phylogenetic diversity, and \code{'FD'} = Functional diversity.}

\item{q}{a numerical vector specifying the diversity orders. Default is c(0, 1, 2).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}) with all entries being \code{0} (non-detection) or \code{1} (detection).}

\item{base}{Sample-sized-based rarefaction and extrapolation for gamma and alpha diversity (\code{base = "size"}) or coverage-based rarefaction and extrapolation for gamma, alpha and beta diversity (\code{base = "coverage"}). Default is \code{base = "coverage"}.}

\item{level}{A numerical vector specifying the particular value of sample coverage (between 0 and 1 when \code{base = "coverage"}) or sample size (\code{base = "size"}). \code{level = 1} (\code{base = "coverage"}) means complete coverage (the corresponding diversity represents asymptotic diversity).\cr
If \code{base = "size"} and \code{level = NULL}, then this function computes the gamma and alpha diversity estimates up to double the reference sample size. \cr 
If \code{base = "coverage"} and \code{level = NULL}, then this function computes the gamma and alpha diversity estimates up to one (for \code{q = 1, 2}) or up to the coverage of double the reference sample size (for \code{q = 0});
the corresponding beta diversity is computed up to the same maximum coverage as the alpha diversity.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter \code{0} to skip the bootstrap procedures. Default is \code{20}. If more accurate results are required, set \code{nboot = 100} (or \code{200}).}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{0.95}.}

\item{PDtree}{(required only when \code{diversity = "PD"}), a phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{PDreftime}{(required only when \code{diversity = "PD"}), a numerical value specifying reference time for PD. Default is \code{NULL} (i.e., the age of the root of PDtree).}

\item{PDtype}{(required only when \code{diversity = "PD"}), select PD type: \code{PDtype = "PD"} (effective total branch length) or \code{PDtype = "meanPD"} (effective number of equally divergent lineages). Default is \code{"meanPD"}, where \code{meanPD = PD/tree depth}.}

\item{FDdistM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species in the pooled assemblage.}

\item{FDtype}{(required only when \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_value"} for FD under a specified threshold value, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{(required only when \code{diversity = "FD"} and \code{FDtype = "tau_value"}), a numerical value between 0 and 1 specifying the tau value (threshold level). If \code{NULL} (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled assemblage (i.e., quadratic entropy).}

\item{FDcut_number}{(required only when \code{diversity = "FD"} and \code{FDtype = "AUC"}), a numeric number to split zero to one into several equal-spaced length. Default is 30.}
}
\value{
If \code{base = "coverage"}, return a list of seven matrices with three diversity (gamma, alpha, and beta diversity) and four dissimilarity indices. If \code{base = "size"}, return a list of two matrices with two diversity (gamma and alpha diversity).
}
\description{
\code{iNEXTbeta3D}: Interpolation and extrapolation of beta diversity with order q
}
\examples{
## Taxonomic diversity for abundance data
# Coverage-based
data(Brazil_data)
output1c = iNEXTbeta3D(data = Brazil_data[c('Marim', 'Rebio 2')], diversity = 'TD', datatype = 'abundance', 
                       base = "coverage", nboot = 30, conf = 0.95)
output1c


# Size-based
data(Brazil_data)
output1s = iNEXTbeta3D(data = Brazil_data[c('Marim', 'Rebio 2')], diversity = 'TD', datatype = 'abundance', 
                       base = "size", nboot = 30, conf = 0.95)
output1s


## Taxonomic diversity for incidence data
# Coverage-based
data(SG_rainforests)
data = list("Cuatro_Rios 2005 v.s. 2017" = SG_rainforests$Cuatro_Rios$Year_2005_and_2017,
            "Juan_Enriquez 2005 v.s. 2017" = SG_rainforests$Juan_Enriquez$Year_2005_and_2017)
output2c = iNEXTbeta3D(data = data, diversity = 'TD', datatype = 'incidence_raw', 
                       level = NULL, nboot = 10, conf = 0.95)
output2c


# Size-based
data(SG_rainforests)
data = list("Cuatro_Rios 2005 v.s. 2017" = SG_rainforests$Cuatro_Rios$Year_2005_and_2017,
            "Juan_Enriquez 2005 v.s. 2017" = SG_rainforests$Juan_Enriquez$Year_2005_and_2017)
output2s = iNEXTbeta3D(data = data, diversity = 'TD', datatype = 'incidence_raw', 
                       base = "size", level = NULL, nboot = 10, conf = 0.95)
output2s


## Phylogenetic diversity for abundance data
# Coverage-based
data(Brazil_PDFD_data)
data(Brazil_tree)
output3c = iNEXTbeta3D(data = Brazil_PDFD_data[c('Marim', 'Rebio 2')], diversity = 'PD', datatype = 'abundance', 
                       base = "coverage", nboot = 10, conf = 0.95, 
                       PDtree = Brazil_tree, PDreftime = NULL, PDtype = 'PD')
output3c

# Size-based
data(Brazil_PDFD_data)
data(Brazil_tree)
output3s = iNEXTbeta3D(data = Brazil_PDFD_data[c('Marim', 'Rebio 2')], diversity = 'PD', datatype = 'abundance', 
                       base = "size", nboot = 20, conf = 0.95, 
                       PDtree = Brazil_tree, PDreftime = NULL, PDtype = 'PD')
output3s


## Phylogenetic diversity for incidence data
data(beetle_inc)
data(beetle_tree)
output4 = iNEXTbeta3D(data = beetle_inc, diversity = 'PD', datatype = 'incidence_raw', 
                      level = seq(0.8, 1, 0.05), nboot = 10, conf = 0.95, 
                      PDtree = beetle_tree, PDreftime = NULL, PDtype = 'PD')
output4


## Functional diversity for abundance data under single threshold
# Coverage-based
data(Brazil_PDFD_data)
data(Brazil_distM)
output5c = iNEXTbeta3D(data = Brazil_PDFD_data[c('Marim', 'Rebio 2')], diversity = 'FD', datatype = 'abundance', 
                       base = "coverage", nboot = 30, conf = 0.95, 
                       FDdistM = Brazil_distM, FDtype = 'tau_value', FDtau = NULL)
output5c

# Size-based
data(Brazil_PDFD_data)
data(Brazil_distM)
output5s = iNEXTbeta3D(data = Brazil_PDFD_data[c('Marim', 'Rebio 2')], diversity = 'FD', datatype = 'abundance', 
                       base = "size", nboot = 30, conf = 0.95, 
                       FDdistM = Brazil_distM, FDtype = 'tau_value', FDtau = NULL)
output5s


## Functional diversity for incidence data under single threshold
data(beetle_inc)
data(beetle_distM)
output6 = iNEXTbeta3D(data = beetle_inc, diversity = 'FD', datatype = 'incidence_raw', 
                      level = seq(0.8, 1, 0.05), nboot = 30, conf = 0.95, 
                      FDdistM = beetle_distM, FDtype = 'tau_value', FDtau = NULL)
output6


## Functional diversity for abundance data with thresholds integrating from 0 to 1
# Coverage-based
data(Brazil_PDFD_data)
data(Brazil_distM)
output7c = iNEXTbeta3D(data = Brazil_PDFD_data[c('Marim', 'Rebio 2')], diversity = 'FD', datatype = 'abundance', 
                       base = "coverage", nboot = 0, conf = 0.95, 
                       FDdistM = Brazil_distM, FDtype = 'AUC', FDcut_number = 30)
output7c


# Size-based
data(Brazil_PDFD_data)
data(Brazil_distM)
output7s = iNEXTbeta3D(data = Brazil_PDFD_data[c('Marim', 'Rebio 2')], diversity = 'FD', datatype = 'abundance', 
                       base = "size", nboot = 10, conf = 0.95, 
                       FDdistM = Brazil_distM, FDtype = 'AUC', FDcut_number = 30)
output7s


## Functional diversity for incidence data with thresholds integrating from 0 to 1
data(beetle_inc)
data(beetle_distM)
output8 = iNEXTbeta3D(data = beetle_inc, diversity = 'FD', datatype = 'incidence_raw', 
                      level = seq(0.8, 1, 0.05), nboot = 10, conf = 0.95, 
                      FDdistM = beetle_distM, FDtype = 'AUC', FDcut_number = 30)
output8
}
