% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{iNEXTbeta3D}
\alias{iNEXTbeta3D}
\title{iNterpolation and EXTrapolation for beta diversity}
\usage{
iNEXTbeta3D(
  data,
  diversity = "TD",
  q = c(0, 1, 2),
  datatype = "abundance",
  base = "coverage",
  level = NULL,
  nboot = 10,
  conf = 0.95,
  PDtree = NULL,
  PDreftime = NULL,
  PDtype = "meanPD",
  FDdistM = NULL,
  FDtype = "AUC",
  FDtau = NULL,
  FDcut_number = 30
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a \code{matrix/data.frame} (species by assemblages), or a \code{list} of \code{matrices/data.frames}, each matrix represents species-by-assemblages abundance matrix.\cr
(b) For \code{datatype = "incidence_raw"}, data can be input as a \code{list} (a dataset) with several \code{lists} (assemblages) of \code{matrices/data.frames}, each matrix represents species-by-sampling units.}

\item{diversity}{selection of diversity type: \code{'TD'} = Taxonomic diversity, \code{'PD'} = Phylogenetic diversity, and \code{'FD'} = Functional diversity.}

\item{q}{a numerical vector specifying the diversity orders. Default is c(0, 1, 2).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}) with all entries being \code{0} (non-detection) or \code{1} (detection).}

\item{base}{Sample-sized-based rarefaction and extrapolation for gamma and alpha diversity (\code{base = "size"}) or coverage-based rarefaction and extrapolation for gamma, alpha and beta diversity (\code{base = "coverage"}). Default is \code{base = "coverage"}.}

\item{level}{A numerical vector specifying the particular value of sample coverage (between 0 and 1 when \code{base = "coverage"}) or sample size (\code{base = "size"}). \code{level = 1} (\code{base = "coverage"}) means complete coverage (the corresponding diversity represents asymptotic diversity).\cr
If \code{base = "size"} and \code{level = NULL}, then this function computes the gamma and alpha diversity estimates up to double the reference sample size. \cr 
If \code{base = "coverage"} and \code{level = NULL}, then this function computes the gamma and alpha diversity estimates up to one (for \code{q = 1, 2}) or up to the coverage of double the reference sample size (for \code{q = 0});
the corresponding beta diversity is computed up to the same maximum coverage as the alpha diversity.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter \code{0} to skip the bootstrap procedures. Default is \code{10}. If more accurate results are required, set \code{nboot = 100} (or \code{200}).}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{0.95}.}

\item{PDtree}{(required only when \code{diversity = "PD"}), a phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{PDreftime}{(required only when \code{diversity = "PD"}), a numerical value specifying reference time for PD. Default is \code{NULL} (i.e., the age of the root of PDtree).}

\item{PDtype}{(required only when \code{diversity = "PD"}), select PD type: \code{PDtype = "PD"} (effective total branch length) or \code{PDtype = "meanPD"} (effective number of equally divergent lineages). Default is \code{"meanPD"}, where \code{meanPD = PD/tree depth}.}

\item{FDdistM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species in the pooled assemblage.}

\item{FDtype}{(required only when \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_value"} for FD under a specified threshold value, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{(required only when \code{diversity = "FD"} and \code{FDtype = "tau_value"}), a numerical value between 0 and 1 specifying the tau value (threshold level). If \code{NULL} (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled assemblage (i.e., quadratic entropy).}

\item{FDcut_number}{(required only when \code{diversity = "FD"} and \code{FDtype = "AUC"}), a numeric number to split zero to one into several equal-spaced length. Default is 30.}
}
\value{
If \code{base = "coverage"}, return a list of seven matrices with three diversity (gamma, alpha, and beta diversity) and four dissimilarity measures. If \code{base = "size"}, return a list of two matrices with two diversity (gamma and alpha diversity). The argument of each column see below:\cr\cr
'Dataset' = the datasets name.\cr\cr
'Order.q' = the diversity order of q.\cr\cr
'SC' = the target standardized coverage value. The observed coverage and extrapolation limit for beta diversity are defined the same as those for alpha diversity. For q = 0, the extrapolation can be extended to a maximum coverage value C(2n, alpha) = coverage value of twice the alpha reference sample size; for q = 1 and 2, target coverage can be extended to 1 (complete coverage) if data are not sparse. \cr\cr
'Size' = the corresponding sample size for the standardized coverage value. \cr\cr
'Alpha/Beta/Gamma/Dissimilarity' = the estimated diversity or dissimilarity of order q for the target coverage value. The estimate for complete coverage (or size = infinity) represents the estimated asymptotic diversity. \cr\cr
'Method' = Rarefaction, Observed, or Extrapolation, depending on whether the target coverage is less than, equal to, or greater than the coverage of the reference sample. (For beta diversity, observed coverage is defined as the coverage of the alpha reference sample).\cr\cr
's.e.' = standard error of diversity estimate.\cr\cr
'LCL', 'UCL' = the bootstrap lower and upper confidence limits for the diversity of order q at the specified level (with a default value of 0.95).\cr\cr
'Diversity' = "TD" (taxonomic diversity), "PD" (phylogenetic diversity of effective total branch length), "meanPD" (phylogenetic diversity of effective number of equally divergent lineages), "FD_tau" (functional diversity under a single tau), "FD_AUC" (functional diversity by integrating all threshold values between zero and one.\cr\cr
'Tau' = the threshold of functional distinctiveness between any two species.\cr
}
\description{
\code{iNEXTbeta3D}: Interpolation and extrapolation of beta diversity with order q
}
\examples{
## Taxonomic diversity for abundance data
# Coverage-based
data(Brazil_rainforests)
output1c = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'TD', 
                       datatype = 'abundance', base = "coverage", nboot = 30, conf = 0.95)
output1c


# Size-based
data(Brazil_rainforests)
output1s = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'TD', 
                       datatype = 'abundance', base = "size", nboot = 30, conf = 0.95)
output1s


## Phylogenetic diversity for abundance data
# Coverage-based
data(Brazil_rainforests)
data(Brazil_tree)
output2c = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'PD', 
                       datatype = 'abundance', base = "coverage", nboot = 10, conf = 0.95, 
                       PDtree = Brazil_tree, PDreftime = NULL, PDtype = 'PD')
output2c

# Size-based
data(Brazil_rainforests)
data(Brazil_tree)
output2s = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'PD', 
                       datatype = 'abundance', base = "size", nboot = 20, conf = 0.95, 
                       PDtree = Brazil_tree, PDreftime = NULL, PDtype = 'PD')
output2s


## Functional diversity for abundance data under single threshold
# Coverage-based
data(Brazil_rainforests)
data(Brazil_distM)
output3c = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'FD', 
                       datatype = 'abundance', base = "coverage", nboot = 30, conf = 0.95, 
                       FDdistM = Brazil_distM, FDtype = 'tau_value', FDtau = NULL)
output3c

# Size-based
data(Brazil_rainforests)
data(Brazil_distM)
output3s = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'FD', 
                       datatype = 'abundance', base = "size", nboot = 30, conf = 0.95, 
                       FDdistM = Brazil_distM, FDtype = 'tau_value', FDtau = NULL)
output3s


## Functional diversity for abundance data with thresholds integrating from 0 to 1
# Coverage-based
data(Brazil_rainforests)
data(Brazil_distM)
output4c = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'FD', 
                       datatype = 'abundance', base = "coverage", nboot = 0, conf = 0.95, 
                       FDdistM = Brazil_distM, FDtype = 'AUC', FDcut_number = 30)
output4c


# Size-based
data(Brazil_rainforests)
data(Brazil_distM)
output4s = iNEXTbeta3D(data = Brazil_rainforests, diversity = 'FD', 
                       datatype = 'abundance', base = "size", nboot = 10, conf = 0.95, 
                       FDdistM = Brazil_distM, FDtype = 'AUC', FDcut_number = 30)
output4s


## Taxonomic diversity for incidence data
# Coverage-based
data(Second_growth_forests)
output5c = iNEXTbeta3D(data = Second_growth_forests, diversity = 'TD', datatype = 'incidence_raw', 
                       level = NULL, nboot = 10, conf = 0.95)
output5c


# Size-based
data(Second_growth_forests)
output5s = iNEXTbeta3D(data = Second_growth_forests, diversity = 'TD', datatype = 'incidence_raw', 
                       base = "size", level = NULL, nboot = 10, conf = 0.95)
output5s


}
