% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{print.iNEXTbeta3D}
\alias{print.iNEXTbeta3D}
\title{Printing \code{iNEXTbeta3D} object}
\usage{
\method{print}{iNEXTbeta3D}(x, ...)
}
\arguments{
\item{x}{an \code{iNEXTbeta3D} object computed by \code{iNEXTbeta3D}.}

\item{...}{additional arguments.}
}
\value{
a list of multiple objects (see \code{iNEXTbeta3D} for more details) with simplified outputs.
}
\description{
\code{print.iNEXTbeta3D}: Print method for objects inheriting from class \code{"iNEXTbeta3D"}
}
